Moved validation functions to a separate file
